--- mlflow/_assertions/__init__.py
@@ -1,3 +1,0 @@
-from mlflow._assertions.decorator import test
-
-__all__ = ["test"]
