+++ mlflow/pytest/__init__.py
@@ -1,0 +1,3 @@
+from mlflow.pytest.decorator import test
+
+__all__ = ["test"]
